+++ XF/Service/Post/OwnerChanger.php
@@ -149,6 +149,8 @@ class OwnerChanger extends AbstractOwnerChanger
     }
 
     /**
+     * @version 2.0.14
+     *
      * @param Entity|ExtendedPostEntity $content
      *
      * @throws \Exception
@@ -161,6 +163,13 @@ class OwnerChanger extends AbstractOwnerChanger
         {
             $threadRepo = $this->getThreadRepo();
             $threadRepo->rebuildThreadPostPositions($content->thread_id);
+        }
+
+        $thread = $content->Thread;
+        if ($thread)
+        {
+            $thread->rebuildCounters();
+            $thread->saveIfChanged();
         }
     }